microservices/geofence/requirements.txt
grpcio==1.48.1
grpcio-tools==1.48.1
requests==2.26.0
protobuf==3.20.0